MODEL: Addition of Quebec load database
Addition of the Quebec load database.
Unlike other databases, this load database is not directly sourced
from Quebec Hydro.  The load database relies on data extracted from
two documents and so should be viewed as less accurate and more
coarse than others.  It is, however, the best available starting
point.