MODEL: Integrate Asset List into AB Spreadsheet support
Added asset list, with associated fuel and capacity info,
into Alberta load data gathering.  This adds the following
output files:
- gen_db_AB_2019 - list of assets, by fuel, for 2019
- gen_pv_AB_actual.txt - PV Solar generation actual figures
- gen_wind_AB_actual.txt - Wind generation actual figures

As part of this development, data structure and procedure
for converting a string into a fuel type was moved to
Common/common_defs.py.  Test code was moded as well.

Additionally, generator_data_gathering changed to use
a generator fuel type column, where available, instead
of the fuel for the table heading.  This has improved
the accuracy of a few of the gen_db_*.txt files.