MODEL: Make hourly_gen_file more generic
Preparation for allowing demand_file to be
a refinement of hourly_gen_file.

Rename hourly_gen_file to hourly_mw_file, to
make it more generic.  Allow hourly_gen_file
to handle different numbers of tokens in the
command line, based on the number of tokens
in the header.  Always keep the file lines
in a list of lines, with the original file
line and line number as the first two tokens.
Use more generic "mw" based names for methods.